+++ src/math/CorrelationCalculator.java
@@ -5,12 +5,12 @@ import list.ValuePair;
 
 public class CorrelationCalculator {
 	
-	public double getCorrelation(PairList<Number> values) {
-		Double x = 0d;
-		Double y = 0d;
-		Double xy = 0d;
-		Double xx = 0d;
-		Double yy = 0d;
+	public float getCorrelation(PairList<Number> values) {
+		Float x = 0f;
+		Float y = 0f;
+		Float xy = 0f;
+		Float xx = 0f;
+		Float yy = 0f;
 		
 		while(values.hasNext()) {
 			@SuppressWarnings("unchecked")
@@ -18,18 +18,18 @@ public class CorrelationCalculator {
 			if(n.getX() == null || n.getY() == null)
 				continue;
 			
-			x += n.getX().doubleValue();
-			y += n.getY().doubleValue();
-			xy += n.getX().doubleValue() * n.getY().doubleValue();
-			xx += n.getX().doubleValue() * n.getX().doubleValue();
-			yy += n.getY().doubleValue() * n.getY().doubleValue();
+			x += n.getX().floatValue();
+			y += n.getY().floatValue();
+			xy += n.getX().floatValue() * n.getY().floatValue();
+			xx += n.getX().floatValue() * n.getX().floatValue();
+			yy += n.getY().floatValue() * n.getY().floatValue();
 		}
 		
-		double top = (values.getActualCount() * xy) - (x * y);
-		double bottom1 = (values.getActualCount() * xx) - (x * x);
-		double bottom2 = (values.getActualCount() * yy) - (y * y);
+		float top = (values.getActualCount() * xy) - (x * y);
+		float bottom1 = (values.getActualCount() * xx) - (x * x);
+		float bottom2 = (values.getActualCount() * yy) - (y * y);
 		
-		double result = top / java.lang.Math.sqrt(bottom1 * bottom2);
+		float result = (float) (top / java.lang.Math.sqrt(bottom1 * bottom2));
 		
 //		System.out.println(values.getActualCount()+" * "+xy+" - "+x+" * "+y);
 //		System.out.println("("+values.getActualCount()+" * "+xx+" - "+x+"^2) * ("+values.getActualCount()+" * "+yy+" - "+y+"^2)");
