+++ src/math/SignificanceCalculator.java
@@ -11,17 +11,17 @@ public class SignificanceCalculator {
 		correlation = new CorrelationCalculator();
 	}
 	
-	public double getSignificance(PairList<Number> values) {
+	public float getSignificance(PairList<Number> values) {
 		if(values.getActualCount() < 2)
 			return 0;
 		
-		double c = correlation.getCorrelation(values);
+		float c = correlation.getCorrelation(values);
 		
-		double top1 = java.lang.Math.abs(c);
-		double top2 = java.lang.Math.sqrt(values.getActualCount() - 2);
-		double bottom = java.lang.Math.sqrt(1 - (c * c));
+		float top1 = java.lang.Math.abs(c);
+		float top2 = (float) java.lang.Math.sqrt(values.getActualCount() - 2);
+		float bottom = (float) java.lang.Math.sqrt(1 - (c * c));
 		
-		double t = (top1 * top2) / bottom;
+		float t = (top1 * top2) / bottom;
 		System.out.println("t: "+t);
 		
 		TDistribution tDistribution = new TDistribution(values.getActualCount() - 2);
